[#828] Memoized Headers File
Fix bugs with export, write out memoized headers to file, read memoized
headers from file before export. Update specs to reflect changes.
