+++ memoized_mesh.txt
@@ -1,0 +1,1 @@
+{"Anti-Infective Agents"=>"D000890", " Local"=>"D000772", "Thinking"=>"D013850", "Problem Solving"=>"D011340", "Morals"=>"D009014", "Recreation"=>"D000068316", "Water Supply"=>"D000082463", "Vision"=>"D000077776", "Art"=>"D000066511", "Hemiptera"=>"D006430", "Oligochaeta"=>"D009835", "Anesthetics"=>"D000777", "Surgery"=>"D000067368", " Oral"=>"D000075067", "Travel"=>"D000076082", "Scarlet Fever"=>"D012541", "Rest"=>"D000071557", "Industry"=>"D000066388", "Inflammation"=>"D007249", "Social Class"=>"D012923", "Fossils"=>"D005580", "Wood"=>"D014934", "Hominidae"=>"D015186", "Human Activities"=>"D006802", "Dreams"=>"D004325", "Vocational Education"=>"D014829", "Education"=>"D000071195", "Cassia"=>"D002366", "Police"=>"D000071823", "Observation"=>"D000067289", "Civilization"=>"D002962", "Suppuration"=>"D013492", "Oral Surgical Procedures"=>"D013516", "Autopsy"=>"D001344", "Salicylic Acid"=>"D000636", "Students"=>"D013334", " Dental"=>"D000077824", "Food Chain"=>"D020387", "Fungi"=>"D000077336", "Typhoid Fever"=>"D010284", "Trees"=>"D003663", "Solutions"=>"D000077324", "Weather"=>"D000080165", "Wind"=>"D010046", "Rain"=>"D000066908", "Biochemistry"=>"D001671", "Tooth"=>"D002607", "Dental Models"=>"D003765", "Histological Techniques"=>"D006652", "Dental Clinics"=>"D003740", "Dentistry"=>"D003149", "Editorial"=>"D016421", "Neurology"=>"D009462", "Pediatrics"=>"D010372", "Nervous System Diseases"=>"D001342", "Child Development"=>"D002657", "Brain Diseases"=>"D001927", "Neurosurgery"=>"D009493", "Child"=>"D000067454", "Infant"=>"D000068104", "Headache Disorders"=>"D020773", "Sleep Disorders"=>"D020178", "Movement Disorders"=>"D009069", "Schools, Medical"=>"D012577", "History"=>"D000071421", "Libraries, Medical"=>"D007993", "Chicago"=>"D002641", "Integrative Medicine"=>"D055048", "Hispanic Americans"=>"D006630", "Surveys and Questionnaires"=>"D011795", "Complementary Therapies"=>"D000529", "Vocabulary, Controlled"=>"D018875"}
