+++ spec/models/invenio_rdm_record_converter_spec.rb
@@ -4,6 +4,10 @@ RSpec.describe InvenioRdmRecordConverter do
   let(:user) { FactoryGirl.create(:user, username: "usr1234", formal_name: "Tester, Mock", orcid: "https://orcid.org/1234-5678-9123-4567", \
                                     display_name: 'Mock Tester') }
   let(:assistant) { FactoryGirl.create(:user, username: "ast9876") }
+  let(:mesh_term) { "Vocabulary, Controlled" }
+  let(:expected_mesh_pid) { "D018875" }
+  let(:lcsh_term) { "Semantic Web" }
+  let(:expected_lcsh_pid) { "sh2002000569" }
   let(:generic_file) {
     make_generic_file(
       user,
@@ -176,140 +180,19 @@ RSpec.describe InvenioRdmRecordConverter do
     end
   end
 
-  # TODO move to separate spec, something like spec/models/header_lookup_spec.rb
-  let(:mesh_term) { "Vocabulary, Controlled" }
-  let(:mesh_query_url) do
-    "https://id.nlm.nih.gov/mesh/sparql?format=JSON&limit=10&inference=true&query=PREFIX%20rdfs%3A%20%3Chttp%3A%2F%2Fwww"\
-    ".w3.org%2F2000%2F01%2Frdf-schema%23%3E%0D%0APREFIX%20meshv%3A%20%3Chttp%3A%2F%2Fid.nlm.nih.gov%2Fmesh%2Fvocab%23%3E"\
-    "%0D%0APREFIX%20mesh2018%3A%20%3Chttp%3A%2F%2Fid.nlm.nih.gov%2Fmesh%3E%0D%0A%0D%0ASELECT%20%3Fd%20%3FdName%0D%0AFROM"\
-    "%20%3Chttp%3A%2F%2Fid.nlm.nih.gov%2Fmesh%3E%0D%0AWHERE%20%7B%0D%0A%20%20%3Fd%20a%20meshv%3ADescriptor%20.%0D%0A%20%"\
-    "20%3Fd%20rdfs%3Alabel%20%3FdName%0D%0A%20%20FILTER(REGEX(%3FdName%2C%27Vocabulary, Controlled%27%2C%20%27i%27))%20%0"\
-    "D%0A%7D%20%0D%0AORDER%20BY%20%3Fd%20%0D%0A"
-  end
-  let(:mesh_api_response) do
-    "{
-      \"head\": {
-        \"vars\": [ \"d\" , \"dName\" ]
-      } ,
-      \"results\": {
-          \"bindings\": [
-          {
-            \"d\": { \"type\": \"uri\" , \"value\": \"http://id.nlm.nih.gov/mesh/D018875\" } ,
-            \"dName\": { \"type\": \"literal\" , \"xml:lang\": \"en\" , \"value\": \"Vocabulary, Controlled\" }
-          }
-        ]
-      }
-    }"
-  end
-  let(:empty_mesh_api_response) do
-    "{
-     \"head\": {
-       \"vars\": [ \"d\" , \"dName\" ]
-       } ,
-       \"results\": {
-           \"bindings\": [
-           ]
-       }
-     }"
-  end
-  let(:expected_mesh_pid) { "D018875" }
-  let(:expected_memoized_mesh) { {mesh_term=>expected_mesh_pid} }
-  let(:blank_mesh_term) { "mesh term does not exist" }
-  let(:expected_failed_mesh) { "#{blank_mesh_term} - MESH" }
+  let(:blank_text_field) { [""] }
 
-  describe "#mesh_term_pid_lookup" do
-    context 'search returns no values' do
-      before do
-        allow(HTTParty).to receive(:get).and_return(mesh_api_response)
-        @mesh_pid = converter.mesh_term_pid_lookup(mesh_term)
-      end
-
-      it 'calls api with correct term' do
-        expect(HTTParty).to have_received(:get).with(mesh_query_url)
-      end
-
-      it 'returns PID for mesh term' do
-        expect(@mesh_pid).to eq(expected_mesh_pid)
-      end
-
-      it 'memoizes the result on success' do
-        expect(converter.send(:memoized_mesh_lookups)).to eq(expected_memoized_mesh)
+  context 'when text field is empty' do
+    describe "#additional_titles" do
+      it 'returns empty array' do
+        expect(subject.send(:additional_titles, blank_text_field)).to eq([])
       end
     end
 
-    context 'search returns no values' do
-      before do
-        allow(HTTParty).to receive(:get).and_return(empty_mesh_api_response)
-        @mesh_pid = converter.mesh_term_pid_lookup(blank_mesh_term)
+    describe "#additional_descriptions" do
+      it 'returns empty array' do
+        expect(subject.send(:additional_descriptions, blank_text_field)).to eq([])
       end
-
-      it 'returns N/A for PID for mesh term' do
-        expect(@mesh_pid).to eq(nil)
-      end
-    end
-  end
-
-  let(:lcsh_term) { "Semantic Web" }
-  let(:lcsh_query_url) { "http://id.loc.gov/authorities/subjects/suggest/?q=*Semantic*Web*" }
-  let(:lcsh_api_response) do
-    """\
-    [\"*Semantic*Web*\",[\"Semantic Web\",\"Semantic Web--Congresses\"],[\"1 result\",\"1 result\"],[\"http://id.loc.go\
-    v/authorities/subjects/sh2002000569\",\"http://id.loc.gov/authorities/subjects/sh2010112582\"]]
-    """
-  end
-  let(:empty_lcsh_api_response) { "[\"*this*is*not*a*real*search*\",[],[],[]]" }
-  let(:expected_lcsh_pid) { "sh2002000569" }
-  let(:expected_memoized_lcsh) { {lcsh_term=>expected_lcsh_pid} }
-  let(:blank_lcsh_term) { "lcsh term does not exist" }
-  let(:expected_failed_lcsh) { "#{blank_lcsh_term} - LCSH" }
-
-  describe "#lcsh_term_pid_lookup" do
-    context 'search returns values' do
-      before do
-        allow(HTTParty).to receive(:get).and_return(lcsh_api_response)
-        @lcsh_pids = converter.lcsh_term_pid_lookup(lcsh_term)
-      end
-
-      it 'calls api with correct term' do
-        expect(HTTParty).to have_received(:get).with(lcsh_query_url)
-      end
-
-      # TODO figure out what to do with terms that have multiple PIDs returned from query
-      it 'returns PID for lcsh term' do
-        expect(@lcsh_pids).to eq(expected_lcsh_pid)
-      end
-
-      it 'memoizes the result on success' do
-        expect(converter.send(:memoized_lcsh_lookups)).to eq(expected_memoized_lcsh)
-      end
-    end
-
-    context 'search returns no values' do
-      before do
-        allow(HTTParty).to receive(:get).and_return(empty_lcsh_api_response)
-        @lcsh_pids = converter.lcsh_term_pid_lookup(lcsh_term)
-      end
-
-      it 'returns nil for PID for lcsh term' do
-        expect(@lcsh_pids).to eq(nil)
-      end
-    end
-  end
-
-  let(:no_pid_mesh_subject) { [{subject: "Fake Mesh Term: DigitalHub field mesh"}] }
-  let(:no_pid_lcsh_subject) { [{subject: "Fake Lcsh Term: DigitalHub field lcsh"}] }
-
-  describe "#subjects_for_scheme" do
-    before do
-      allow(subject).to receive(:pid_lookup_by_scheme).and_return(nil)
-    end
-
-    it 'returns properly formatted subject for lcsh subject with no pid' do
-      expect(subject.send(:subjects_for_scheme, ["Fake Lcsh Term"], :lcsh)).to eq(no_pid_lcsh_subject)
-    end
-
-    it 'returns properly formatted subject for mesh subject with no pid' do
-      expect(subject.send(:subjects_for_scheme, ["Fake Mesh Term"], :mesh)).to eq(no_pid_mesh_subject)
     end
   end
 end
